AI_gen_workloads/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
groq==0.30.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.69
langchain-groq==0.3.6
langchain-openai==0.3.28
langchain-text-splitters==0.3.8
langsmith==0.4.8
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.97.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.11.0
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pydantic==2.11.7
pydantic_core==2.33.2
pyperclip==1.9.0
PySocks==1.7.1
pytesseract==0.3.13
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
psycopg2-binary >= 2.5